requirements.txt
streamlit
torch
torchvision
transformers
pillow
requests
matplotlib
nltk
rouge-score
numpy
pandas
plotly